enforce_type in util
